fix: missing go back button on squads page
